Move constraints to domain package
