+++ optaplanner-quickstarts/quarkus-facility-location/src/main/java/org/acme/facilitylocation/domain/FacilityLocationConstraintConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.acme.facilitylocation.solver;
+package org.acme.facilitylocation.domain;
 
 import org.optaplanner.core.api.domain.constraintweight.ConstraintConfiguration;
 import org.optaplanner.core.api.domain.constraintweight.ConstraintWeight;
@@ -23,9 +23,9 @@ import org.optaplanner.core.api.score.buildin.hardsoftlong.HardSoftLongScore;
 @ConstraintConfiguration
 public class FacilityLocationConstraintConfiguration {
 
-    public static final String FACILITY_CAPACITY = "facility capacity";
-    public static final String FACILITY_SETUP_COST = "facility setup cost";
-    public static final String DISTANCE_FROM_FACILITY = "distance from facility";
+    static final String FACILITY_CAPACITY = "facility capacity";
+    static final String FACILITY_SETUP_COST = "facility setup cost";
+    static final String DISTANCE_FROM_FACILITY = "distance from facility";
 
     @ConstraintWeight(FACILITY_CAPACITY)
     HardSoftLongScore facilityCapacity = HardSoftLongScore.ofHard(1);
