+++ optaplanner-quickstarts/quarkus-facility-location/src/test/java/org/acme/facilitylocation/domain/FacilityLocationConstraintProviderTest.java
@@ -14,12 +14,8 @@
  * limitations under the License.
  */
 
-package org.acme.facilitylocation.solver;
+package org.acme.facilitylocation.domain;
 
-import org.acme.facilitylocation.domain.Consumer;
-import org.acme.facilitylocation.domain.Facility;
-import org.acme.facilitylocation.domain.FacilityLocationProblem;
-import org.acme.facilitylocation.domain.Location;
 import org.junit.jupiter.api.Test;
 import org.optaplanner.test.api.score.stream.ConstraintVerifier;
